net/mlx5: fix IPv6 DSCP offset in HWS sync API
The processing of field IPv6 DSCP via the synchronous flow API is
different between software and hardware steering engines(HWS). The bit
shift handling for IPv6 DSCP was not handled in HWS, resulting in
incorrect data in the field.

To resolve this, bit shift handling should be added to HWS.

Fixes: ec1e7a5ceb69 ("net/mlx5: update IPv6 traffic class modification")
Cc: stable@dpdk.org
Signed-off-by: Gavin Li <gavinl@nvidia.com>
Acked-by: Bing Zhao <bingz@nvidia.com>