net/idpf: fix missing offload from conversion function
The TCP_TSO tx offload was missing from the conversion function. This
was effectively taking the request for that offload out of consideration
when selecting the tx function. As a result a vector path which does not
support TSO could be chosen when that offload was requested, when
instead the scalar path should be chosen which does support that
offload. Fix this by adding the TSO offload to the conversion function.

Fixes: c008a5e740bd ("common/idpf: add queue setup/release")
Cc: stable@dpdk.org

Signed-off-by: Ciara Loftus <ciara.loftus@intel.com>
Acked-by: Praveen Shetty <praveen.shetty@intel.com>